ExpressionMap

MultivariateInterpolation.ExpressionMap — Type

ExpressionMap{T <: AbstractAlgebraicVariety}

A data type that represents a rational map $\varphi \colon X \dashrightarrow \mathbb{C}^m$.

Constructor

ExpressionMap(
    domain::AbstractAlgebraicVariety;
    expressions::Pair{<:AbstractArray, <:AbstractArray}=Pair(Variable[], Expression[]),
    projection::AbstractArray=Variable[]
)

Examples

julia> @var R[1:2,1:2] t[1:2] s[1:2]
(Variable[R₁₋₁ R₁₋₂; R₂₋₁ R₂₋₂], Variable[t₁, t₂], Variable[s₁, s₂])

julia> X = AlgebraicVariety([R'*R-I, det(R)-1]; variables=[R, t]);

julia> φ = ExpressionMap(X; expressions=Pair(s, R*t), projection=t)
ExpressionMap: ℂ⁶ ⊇ X - - > ℂ⁴
 domain:
  AlgebraicVariety X ⊂ ℂ⁶
   6 variables: R₁₋₁, R₂₋₁, R₁₋₂, R₂₋₂, t₁, t₂
   5 expressions: 
    -1 + R₁₋₁^2 + R₂₋₁^2
    R₁₋₁*R₁₋₂ + R₂₋₁*R₂₋₂
    R₁₋₁*R₁₋₂ + R₂₋₁*R₂₋₂
    -1 + R₁₋₂^2 + R₂₋₂^2
    -1 + R₁₋₁*R₂₋₂ - R₁₋₂*R₂₋₁
 action:
  s₁ = t₁*R₁₋₁ + t₂*R₁₋₂
  s₂ = t₁*R₂₋₁ + t₂*R₂₋₂
  projection to t₁, t₂

MultivariateInterpolation.domain — Method

domain(φ::ExpressionMap{T<:AbstractAlgebraicVariety}) -> T

Return the domain of φ.

Examples

julia> domain(φ)
AlgebraicVariety X ⊂ ℂ⁶
 6 variables: R₁₋₁, R₂₋₁, R₁₋₂, R₂₋₂, t₁, t₂
 5 expressions: 
  -1 + R₁₋₁^2 + R₂₋₁^2
  R₁₋₁*R₁₋₂ + R₂₋₁*R₂₋₂
  R₁₋₁*R₁₋₂ + R₂₋₁*R₂₋₂
  -1 + R₁₋₂^2 + R₂₋₂^2
  -1 + R₁₋₁*R₂₋₂ - R₁₋₂*R₂₋₁

MultivariateInterpolation.domain_vars — Method

domain_vars(φ::ExpressionMap) -> Vector{Variable}

Return the variables of the domain of φ.

Examples

julia> domain_vars(φ)
6-element Vector{Variable}:
 R₁₋₁
 R₂₋₁
 R₁₋₂
 R₂₋₂
   t₁
   t₂

MultivariateInterpolation.ndomain_vars — Method

ndomain_vars(φ::ExpressionMap) -> Int

Return the number of variables in the domain of φ.

Examples

julia> ndomain_vars(φ)
6

MultivariateInterpolation.expression_vars — Method

expression_vars(φ::ExpressionMap) -> Vector{Variable}

Return the expression variables of φ.

Examples

julia> expression_vars(φ)
2-element Vector{Variable}:
 s₁
 s₂

MultivariateInterpolation.nexpression_vars — Method

nexpression_vars(φ::ExpressionMap) -> Int

Return the number of expression variables of φ.

Examples

julia> nexpression_vars(φ)
2

MultivariateInterpolation.projection_vars — Method

projection_vars(φ::ExpressionMap) -> Vector{Variable}

Return the projection variables of φ.

Examples

julia> projection_vars(φ)
2-element Vector{Variable}:
 t₁
 t₂

MultivariateInterpolation.nprojection_vars — Method

nprojection_vars(φ::ExpressionMap) -> Int

Return the number of projection variables of φ.

Examples

julia> nprojection_vars(φ)
2

MultivariateInterpolation.image_vars — Method

image_vars(φ::ExpressionMap) -> Vector{Variable}

Return the concatenated vector of expression and projection variables of φ.

Examples

julia> image_vars(φ)
4-element Vector{Variable}:
 s₁
 s₂
 t₁
 t₂

MultivariateInterpolation.nimage_vars — Method

nimage_vars(φ::ExpressionMap) -> Int

Return the number of image variables of φ.

Examples

julia> nimage_vars(φ)
4

MultivariateInterpolation.variables — Method

variables(φ::ExpressionMap) -> Vector{Variable}

Return the concatenated vector of domain and expression variables of φ.

Examples

julia> variables(φ)
8-element Vector{Variable}:
 R₁₋₁
 R₂₋₁
 R₁₋₂
 R₂₋₂
   t₁
   t₂
   s₁
   s₂

MultivariateInterpolation.nvariables — Method

nvariables(φ::ExpressionMap) -> Int

Return the number of variables of φ.

Examples

julia> nvariables(φ)
8

MultivariateInterpolation.expressions — Method

expressions(φ::ExpressionMap) -> Vector{Expression}

Return the expressions that define φ. Doesn't include the projection variables.

Examples

julia> expressions(φ)
2-element Vector{Expression}:
 t₁*R₁₋₁ + t₂*R₁₋₂
 t₁*R₂₋₁ + t₂*R₂₋₂

MultivariateInterpolation.domain_dimension — Function

domain_dimension(φ::ExpressionMap; <keyword_arguments>) -> Int

Compute the dimension of the domain of $\varphi$.

Examples

julia> domain_dimension(φ)
3

MultivariateInterpolation.image_dimension — Function

image_dimension(φ::ExpressionMap; <keyword_arguments>) -> Int

Compute the dimension of the image of $\varphi$.

Examples

julia> image_dimension(φ)
3

MultivariateInterpolation.is_dominant — Function

is_dominant(φ::ExpressionMap; <keyword_arguments>)

Return true if $\varphi \colon X \dashrightarrow \mathbb{C}^m$ is dominant, i.e. if

\[\mathrm{dim}(\mathrm{im}(\varphi)) = m.\]

Examples

julia> is_dominant(φ)
false